# types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Headless Form Handler",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8001/api"
    }
  ],
  "security": [
    {
      "http": []
    }
  ],
  "paths": {
...
    "/v1/auth/refresh": {
      "post": {
        "operationId": "auth.refresh",
        "description": "The old token is blacklisted. Refreshing is allowed until the\ntoken is older than the configured refresh TTL, and never for a\ntoken revoked by a password change or reset: the new token keeps\nthe original token version, so it is checked against revocation.",
        "summary": "Exchange a current or recently expired token for a new one",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string",
                      "const": "bearer"
                    },
                    "expires_in": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "access_token",
                    "token_type",
                    "expires_in"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/AuthenticationException"
          }
        },
        "security": []
      }
    },
...
    "/v1/auth/forgot-password": {
      "post": {
        "operationId": "auth.forgot-password",
        "summary": "Email a password reset link. The response is the same whether or not the account exists, or\nthe request was throttled, so the endpoint cannot be used to discover registered emails",
        "tags": [
          "PasswordReset"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "const": "If an account exists for that email, a password reset link has been sent."
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationException"
          }
        },
        "security": []
      }
    },
    "/v1/auth/reset-password": {
      "post": {
        "operationId": "auth.reset-password",
        "summary": "Set a new password from an emailed reset token and revoke every existing token",
        "tags": [
          "PasswordReset"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {}
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationException"
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "http": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ForgotPasswordRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest"
      },
...
      "ResetPasswordRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "password_confirmation": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "email",
          "password",
          "password_confirmation"
        ],
        "title": "ResetPasswordRequest"
      },
...
    },
    "responses": {
      "ValidationException": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Errors overview."
                },
                "errors": {
                  "type": "object",
                  "description": "A detailed description of each field that failed validation.",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "required": [
                "message",
                "errors"
              ]
            }
          }
        }
      },
      "AuthenticationException": {
        "description": "Unauthenticated",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "description": "Error overview."
                }
              },
              "required": [
                "message"
              ]
            }
          }
        }
      },
...
    }
  }
}